Add test

+++ django_easy_report/models.py
@@ -24,7 +24,7 @@ class ReportSender(models.Model):
     updated_at = models.DateTimeField(auto_now=True)
     name = models.CharField(max_length=32, unique=True, db_index=True)
     email_from = models.EmailField(
-        null=True, blank=True,
+        blank=True, null=True,
         help_text=_('If have content email must be send when report is completed.')
     )
     size_to_attach = models.PositiveIntegerField(
@@ -33,15 +33,15 @@ class ReportSender(models.Model):
                     'otherwise the file will be send as attached on the email.')
     )
     storage_class_name = models.CharField(
-        max_length=64, null=True, blank=True,
+        max_length=64, blank=True, null=True,
         help_text=_('Class name for for save the report. '
                     'It must be subclass of django.core.files.storage.Storage')
     )
     storage_init_params = models.TextField(
-        blank=True, help_text=_('JSON with init parameters')
+        blank=True, null=True, help_text=_('JSON with init parameters')
     )
 
-    def __str__(self):
+    def __str__(self):  # pragma: no cover
         return self.name
 
     def clean(self):
@@ -65,28 +65,28 @@ class ReportSender(models.Model):
                             self.storage_class_name
                         )
                     })
+                elif 'storage_init_params' not in errors:
+                    try:
+                        self.get_storage()
+                    except TypeError as type_error:
+                        errors.update({
+                            'storage_init_params': str(type_error)
+                        })
+                    except Exception as ex:
+                        errors.update({
+                            'storage_init_params': _('Error creating storage class: "{}"').format(ex)
+                        })
+
             except (ImportError, ValueError):
                 errors.update({
                     'storage_class_name': _('Class "{}" cannot be imported').format(
                         self.storage_class_name
                     )
                 })
-            if 'storage_init_params' not in errors:
-                try:
-                    self.get_storage()
-                except TypeError as type_error:
-                    errors.update({
-                        'storage_init_params': str(type_error)
-                    })
-                except Exception as ex:
-                    errors.update({
-                        'storage_init_params': _('Error creating storage class: "{}"').format(ex)
-                    })
-
         if not any([self.email_from, self.storage_class_name]):
-            errors.update(
-                _('Email or Storage class must be entered')
-            )
+            errors.update({
+                '__all__': _('Email or Storage class must be entered')
+            })
 
         if errors:
             raise ValidationError(errors)
@@ -120,14 +120,14 @@ class ReportGenerator(models.Model):
         help_text=_('Class name for for generate the report. '
                     'It must be subclass of django_easy_report.reports.ReportBaseGenerator')
     )
-    init_params = models.TextField(blank=True, help_text=_('JSON with init parameters'))
+    init_params = models.TextField(blank=True, null=True, help_text=_('JSON with init parameters'))
     sender = models.ForeignKey(ReportSender, on_delete=models.PROTECT)
     permissions = models.CharField(
         max_length=1024, blank=True, null=True,
         help_text=_(
             'Comma separated permission list. Permission formatted as: {}'
         ).format(
-            '&lt;content_type.app_label&gt;.&lt;content_type.model&gt;.&lt;permission.codename&gt;'
+            '&lt;content_type.app_label&gt;.&lt;permission.codename&gt;'
         )
     )
     always_generate = models.BooleanField(
@@ -143,7 +143,7 @@ class ReportGenerator(models.Model):
         help_text=_('If model is deleted, do not remove the file on storage')
     )
 
-    def __str__(self):
+    def __str__(self):  # pragma: no cover
         return self.name
 
     def clean(self):
@@ -158,22 +158,15 @@ class ReportGenerator(models.Model):
                     'init_params': _('Invalid JSON')
                 })
 
-        if self.class_name:
-            try:
-                cls = import_class(self.class_name)
-                if not issubclass(cls, ReportBaseGenerator):
-                    errors.update({
-                        'class_name': _('Invalid class "{}", must be instance of ReportBaseGenerator').format(
-                            self.class_name
-                        )
-                    })
-            except (ImportError, ValueError):
+        try:
+            cls = import_class(self.class_name)
+            if not issubclass(cls, ReportBaseGenerator):
                 errors.update({
-                    'class_name': _('Class "{}" cannot be imported').format(
+                    'class_name': _('Invalid class "{}", must be instance of ReportBaseGenerator').format(
                         self.class_name
                     )
                 })
-            if 'init_params' not in errors:
+            elif 'init_params' not in errors:
                 try:
                     self.get_report()
                 except TypeError as type_error:
@@ -185,13 +178,22 @@ class ReportGenerator(models.Model):
                         'init_params': _('Error creating report class: "{}"').format(ex)
                     })
 
+        except (ImportError, ValueError):
+            errors.update({
+                'class_name': _('Class "{}" cannot be imported').format(
+                    self.class_name
+                )
+            })
+
         if self.permissions:
             perm_errors = []
-            for permission in self.get_permissions_list():
+            for permission in self.get_permissions():
                 try:
-                    app_label, model, codename = permission.split('.')
-                    ct = ContentType.objects.get(app_label=app_label, model=model)
-                    Permission.objects.get(content_type=ct, codename=codename)
+                    app_label, codename = permission.rsplit('.')
+                    ct = ContentType.objects.filter(app_label=app_label)
+                    if not ct.exists():
+                        raise ContentType.DoesNotExist()
+                    Permission.objects.get(content_type__in=ct, codename=codename)
                 except Permission.DoesNotExist:
                     perm_errors.append(_('Unknown code name for permission: "{}"').format(permission))
                 except ContentType.DoesNotExist:
@@ -206,10 +208,10 @@ class ReportGenerator(models.Model):
         if errors:
             raise ValidationError(errors)
 
-    def get_permissions_list(self):
-        permissions = []
+    def get_permissions(self):
+        permissions = set()
         if self.permissions:
-            permissions = [p.strip() for p in self.permissions.split(',')]
+            permissions = set([p.strip() for p in self.permissions.split(',')])
         return permissions
 
     def get_report(self):
@@ -242,22 +244,44 @@ class ReportQuery(models.Model):
     report = models.ForeignKey(ReportGenerator, on_delete=models.PROTECT)
     filename = models.CharField(max_length=32)
     mimetype = models.CharField(max_length=32, default='application/octet-stream')
-    params = models.TextField(blank=True)
+    params = models.TextField(blank=True, null=True)
     params_hash = models.CharField(max_length=128)
     storage_path_location = models.CharField(max_length=512, blank=True, null=True)
 
     class Meta:
         ordering = ('created_at', )
 
+    def clean(self):
+        super(ReportQuery, self).clean()
+
+        if self.params:
+            try:
+                json.loads(self.params)
+            except json.JSONDecodeError:
+                raise ValidationError({
+                    'params': _('Invalid JSON')
+                })
+
     @staticmethod
     def gen_hash(data_dict):
         sha = hashlib.sha1()
-        for key in sorted(data_dict.keys()):
-            sha.update(key.encode())
-            value = data_dict.get(key)
-            if isinstance(value, dict):
-                value = ReportQuery.gen_hash(**value)
-            sha.update(str(value).encode())
+        if data_dict:
+            if (
+                hasattr(data_dict, 'keys') and hasattr(data_dict, 'get') and
+                callable(data_dict.keys) and callable(data_dict.get)
+            ):
+                for key in sorted(data_dict.keys()):
+                    sha.update(key.encode())
+                    value = data_dict.get(key)
+                    if isinstance(value, dict):
+                        value = ReportQuery.gen_hash(value)
+                    elif isinstance(value, models.Model):
+                        value = '{}({})'.format(
+                            value.__class__.__class__, value.pk
+                        )
+                    sha.update(str(value).encode())
+            else:
+                raise TypeError('data_dict must implements keys and get functions')
         return sha.hexdigest()
 
     def get_report(self):
@@ -265,13 +289,17 @@ class ReportQuery(models.Model):
         kwargs = {}
         if self.params:
             kwargs = json.loads(self.params)
-        return report.setup(self, **kwargs)
+        report.setup(self, **kwargs)
+        return report
 
     def get_file(self):
         storage = self.report.sender.get_storage()
+        if not storage or not self.storage_path_location:
+            return
         if not self.report.always_download:
             try:
-                return redirect(storage.url(self.storage_path_location))
+                url = storage.url(self.storage_path_location)
+                return redirect(url)
             except NotImplementedError:
                 pass
         return storage.open(self.storage_path_location, 'r')
@@ -282,7 +310,8 @@ def delete_report_from_storage(sender, instance, **kwargs):
     if instance.storage_path_location and instance.report.preserve_report:
         return
     storage = instance.report.sender.get_storage()
-    storage.delete(instance.storage_path_location)
+    if storage:
+        storage.delete(instance.storage_path_location)
 
 
 class ReportRequester(models.Model):
